# swagger/swagger.yaml
basePath: /v1/
consumes:
- application/json
definitions:
...
  OfTheDayModel:
    properties:
      date:
        description: The date when this author / quote should be 'of the day'
        example: "2020-06-12"
        type: string
        x-go-name: Date
      id:
        description: The id of the author / quote
        example: 1
        format: int64
        type: integer
        x-go-name: Id
      language:
        default: English
        description: The language of this author / quote
        example: icelandic
        type: string
        x-go-name: Language
    type: object
    x-go-name: ofTheDayModel
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
...
  QodViewAPIModel:
    properties:
      authorId:
        description: The author's id
        format: int64
        type: integer
        x-go-name: AuthorId
      date:
        description: The date when this quote was the quote of the day
        example: "2021-06-12T00:00:00Z"
        type: string
        x-go-name: Date
      isIcelandic:
        description: |-
          Whether the quote is in icelandic
          false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote for the day
        example: Float like a butterfly, sting like a bee
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        x-go-name: QuoteId
    type: object
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
...
  TopicViewAPIModel:
    properties:
      authorId:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: AuthorId
      isIcelandic:
        description: Whether or not this quote is in Icelandic or not
        example: false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote
        example: Float like a butterfly, sting like a bee.
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: QuoteId
      topicId:
        description: The topic's id (if topic id / name not supplied this will return
          a zero id)
        example: 10
        format: int64
        type: integer
        x-go-name: TopicId
      topicName:
        description: The topic's name (if topic id / name not supplied this will return
          empty string "")
        example: inspirational
        type: string
        x-go-name: TopicName
    type: object
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
...
host: rxvglshzhl.execute-api.eu-west-1.amazonaws.com
info:
  contact:
    email: skjaldbaka17@gmail.com
    name: Þórður Ágústsson
  description: Quotel API offers a complete feature rich REST API access to all your
    quotes needs. This is the documentation of our quotes API.
  title: quotes-api.
  version: 1.0.0
paths:
...
  /quotes/qod:
    post:
      description: gets the quote of the day
      operationId: GetQuoteOfTheDay
      parameters:
      - description: The structure of the request for getting the author / quote of
          the day
        in: body
        name: Body
        schema:
          properties:
            language:
              default: English
              description: Get the author / quote of the day for the given language
                ("icelandic" or "english")
              example: English
              type: string
              x-go-name: Language
          type: object
      responses:
        "200":
          $ref: '#/responses/qodResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
...
  /quotes/qod/new:
    post:
      description: Sets the quote of the day for the given dates
      operationId: SetQuoteOfTheDay
      parameters:
      - description: The structure of the request for setting the QOD
        in: body
        name: Body
        schema:
          properties:
            qods:
              items:
                $ref: '#/definitions/OfTheDayModel'
              type: array
              x-go-name: Qods
          type: object
      responses:
        "200":
          $ref: '#/responses/successResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
  /quotes/random:
    post:
      description: Get a random quote according to the given parameters
      operationId: GetRandomQuote
      parameters:
      - description: The structure of the request for a random quote
        in: body
        name: Body
        schema:
          properties:
            authorId:
              description: The random quote returned must be from the author with
                the given authorId
              example: 24952
              format: int64
              type: integer
              x-go-name: Authorid
            language:
              description: The random quote returned must be in the given language
              example: English
              type: string
              x-go-name: Language
            searchString:
              description: The random quote returned must contain a match with the
                searchstring
              example: float
              type: string
              x-go-name: SearchString
            topicId:
              description: The random quote returned must be a part of the topic with
                the given topicId
              example: 10
              format: int64
              type: integer
              x-go-name: TopicId
          type: object
      responses:
        "200":
          $ref: '#/responses/topicViewResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "404":
          $ref: '#/responses/notFoundResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
...
produces:
- application/json
responses:
...
  incorrectBodyStructureResponse:
    description: Data structure representing the error response to a wrongly structured
      request body
    schema:
      properties:
        message:
          description: The error message
          example: request body is not structured correctly.
          type: string
          x-go-name: Message
      type: object
...
  internalServerErrorResponse:
    description: Data structure representing the error response to an internal server
      error
    schema:
      properties:
        message:
          description: The error message
          example: Please try again later.
          type: string
          x-go-name: Message
      type: object
...
  notFoundResponse:
    description: Data structure representing the error response to a not found error
    schema:
      properties:
        message:
          description: The error message
          example: No quote exists that matches the given parameters.
          type: string
          x-go-name: Message
      type: object
...
  qodResponse:
    description: Data structure representing the response for the quote of the day
    schema:
      $ref: '#/definitions/QodViewAPIModel'
...
  successResponse:
    description: ""
    schema:
      properties:
        message:
          example: This request was a success
          type: string
          x-go-name: Message
        statusCode:
          description: HTTP status code
          example: 200
          format: int64
          type: integer
          x-go-name: StatusCode
      type: object
  topicViewResponse:
    description: Data structure representing the response for a quote based on a particular
      topic
    schema:
      $ref: '#/definitions/TopicViewAPIModel'
...
schemes:
- https
swagger: "2.0"
...